fix a bunch of rendering leaks and bugs (re: blending) and add an automated test suite for rendering that doesn't require the game data

+++ Source/relive_lib/Renderer/OpenGL3/Batcher.cpp
@@ -114,7 +114,7 @@ template <typename TextureType, typename RenderBatchType, std::size_t kTextureBa
 void Batcher<TextureType, RenderBatchType, kTextureBatchSize>::InsertVertexData(const IRenderer::PsxVertexData* pVertData, s32 count)
 {
     // Push indicies for this data
-    const u16 nextIndex = mIndexBufferIndex;
+    const u32 nextIndex = mIndexBufferIndex;
     const s32 numTriangles = count - 2;
 
     if (numTriangles == 1)
